add delete validation inviteGame

+++ backend/src/game/rewrite/game-rewrite.service.ts
@@ -261,6 +261,7 @@ export class GameReWriteService {
     roomId,
   }: DeleteGameDto): Promise<InviteGame> {
     // dto.roomIdのデータレコードが存在しているかをprisma.findUniqueで検索
+    console.log(roomId);
     const game = await this.prisma.inviteGame.findUnique({
       where: {
         id: roomId,
@@ -279,16 +280,46 @@ export class GameReWriteService {
     // 上記がどちらとも当てはまらない場合は例外（ForbiddenException）
 
     // 正常の場合はprisma.game.deleteで削除
-    const deleteGame = await this.prisma.inviteGame.delete({
-      where: {
-        id: roomId,
-      },
-    });
+    const Player1 = game.player1;
+    console.log(Player1)
+    const Player2 = game.player2;
+    console.log(Player2)
+    let deleteGame1;
+    let deleteGame2;
+
+    if (Player1) {
+      const isPlayer1InviteGame = await this.prisma.inviteGame.findUnique({
+        where: {
+          player1: Player1,
+        },
+      });
+      if (isPlayer1InviteGame) {
+        deleteGame1 = await this.prisma.inviteGame.delete({
+          where: {
+            player1: Player1,
+          },
+        });
+      }
+    }
+    if (Player2) {
+      const isPlayer2InviteGame = await this.prisma.inviteGame.findUnique({
+        where: {
+          player2: Player2,
+        },
+      });
+      if (isPlayer2InviteGame) {
+        deleteGame2 = await this.prisma.inviteGame.delete({
+          where: {
+            player2: Player2,
+          },
+        });
+      }
+    }
 
     // userNameToInviteGameRoomIdからも削除 player1 player2両方とも
     this.userNameToInviteGameRoomId.delete(game.player1);
     this.userNameToInviteGameRoomId.delete(game.player2);
-    return deleteGame;
+    return deleteGame1 || deleteGame2;
   }
 
   getUserNameToRandomGameRoomId() {
